Update transaction.service.js
Added Aggregate Functions for Expenses and Income

In this commit, two new aggregate functions have been added to the service layer to calculate the total expenses and total income for a user. These functions utilize MongoDB's aggregation framework to perform the calculations.

1. `getExpenses`: This function calculates the total expenses for a user. It uses the `$match` stage to filter transactions with `income` set to `false`, and then the `$group` stage to sum the `sum` field for these transactions. The result is returned as an object containing the `expenses` field.

2. `getIncome`: This function calculates the total income for a user. It follows a similar approach to `getExpenses` but filters transactions with `income` set to `true`.

Both functions take the `userId` as an argument and return the total expenses and total income, respectively.

These aggregate functions provide a way to retrieve financial statistics for users and can be used to calculate balance, expenses, and income.