components/InputsForm
Created InputsForm Component

This commit introduces a new component called 'InputsForm'. It is designed to handle input fields for email and password, along with buttons and links for either logging in or registering. The component accepts two props: 'buttonText' and 'formType' to customize the button text and link text based on the purpose of the form (login or register).

The 'InputsForm' component contains a form element with labels for email and password inputs. Depending on the 'formType' prop value, it either provides a link to sign up or sign in, with the corresponding text. This component can be used to create forms for user registration and login.