components/LoginForm
Created LoginForm Component

This commit adds a new component called 'LoginForm'. It utilizes the 'InputsForm' component to create a login form. The 'LoginForm' component simply renders the 'InputsForm' component with the 'buttonText' set to "Login" and 'formType' set to "login". This allows users to input their email and password for logging in. The component is now available for use in the application.