components/PrivateRoute
Created PrivateRoute Component

This commit adds a new component called 'PrivateRoute'. This component is intended to be used for creating protected routes in a React application. It imports the 'Navigate' component from 'react-router-dom' and uses the 'useAuth' hook to determine whether a user is authenticated.

The 'PrivateRoute' component takes two props:
- 'component': This prop is of type 'React.ComponentType' and represents the component that should be rendered if the user is authenticated.
- 'redirectTo' (optional): If provided, this prop specifies the path to which the user will be redirected if they are not authenticated. If not provided, the user will not be redirected.

Inside the component, it checks the user's authentication status using the 'useAuth' hook. If the user is authenticated, it renders the specified component. If the user is not authenticated, it can redirect to the path specified in the 'redirectTo' prop.

This 'PrivateRoute' component can be used to protect certain routes in your application that should only be accessible to authenticated users.