components/RegisterForm
Created RegisterForm Component

This commit adds a new component called 'RegisterForm'. It is designed for user registration and utilizes the 'InputsForm' component. The 'RegisterForm' component simply renders the 'InputsForm' component with the 'Register' button text and sets the 'formType' prop to 'register'. This allows users to fill in their email and password for registration.

The 'RegisterForm' component can be easily incorporated into the registration page of the application, providing a clean and consistent form for user registration.