auth/operations
Added user registration and login async thunks

This commit includes the creation of two asynchronous thunks for user registration and login. These thunks use axios to make API requests to the server and handle responses and errors accordingly. The setAuthHeader function is used to set the Bearer token in Axios headers upon successful login, and clearAuthHeader is used to remove the token from the headers. These thunks are integrated into the authSlice for state management.