auth/selectors
Added selector functions for authentication state

This commit introduces three selector functions for accessing the authentication state in Redux. `selectIsLoggedIn` is used to check if the user is logged in, `selectUser` retrieves the user information, and `selectIsRefreshing` is used to determine if the authentication state is currently refreshing. These selectors can be used to access and utilize authentication-related information in the application.