components/LoginForm
Added LoginForm component

In this commit, a new component called "LoginForm" has been created. The LoginForm component utilizes the "InputsForm" component, which provides input fields for email and password, as well as a submit button.

The "LoginForm" component sets the "buttonText" prop to "Login" and the "formType" prop to "login." This allows it to reuse the input form structure while customizing the button text and form type for login purposes.

This component will be used in the login page of the application, making it easier to maintain and reuse the input form structure for different authentication forms.